+++ angular/lib/src/core/linker/component_factory.dart
@@ -1,7 +1,6 @@
 import 'dart:html';
 
 import 'package:angular/src/core/di.dart' show Injector;
-import 'package:angular/src/di/reflector.dart' show runtimeTypeProvider;
 
 import '../change_detection/change_detection.dart' show ChangeDetectorRef;
 import 'app_view.dart';
@@ -39,10 +38,6 @@ class ComponentRef<C> {
 
   /// The [ChangeDetectorRef] of the Component instance.
   ChangeDetectorRef get changeDetectorRef => _parentView.viewData.ref;
-
-  /// Returns the type of component [C].
-  @Deprecated('No longer supported, and throws an error in most applications')
-  Type get componentType => runtimeTypeProvider(_component);
 
   /// Destroys the component instance and all of the data structures associated
   /// with it.