+++ angular/lib/src/di/reflector.dart
@@ -20,11 +20,6 @@ export '../core/di/opaque_token.dart' show OpaqueToken;
 // already considered the "slow" path, this isn't a regression.
 final _components = <Object, dynamic /*ComponentFactory*/ >{};
 
-/// May be overridden by legacy clients of AngularDart.
-///
-/// Otherwise, `ComponentRef.componentType` always returns throws.
-Type Function(Object instance) runtimeTypeProvider = _nullTypeProvider;
-
 Type _nullTypeProvider(Object _) {
   // In debug mode give a better error message.
   assert(